Deprecate RecordBatch::concat (#2594)
